refactor(stark): move inlined tests + trace.rs test helpers into src/tests/
Part A — trace.rs:
- Move `get_trace_evaluations` (Horner oracle) to tests/trace_test_helpers.rs
- Move `compute_trace_polys_main` to tests/trace_test_helpers.rs as inherent impl
- Widen `compute_frame_evaluation_points` to pub(crate) (needed by helper)
- Remove all #[cfg(test)] items from trace.rs

Part B — inline test modules:
- grinding.rs → tests/grinding_tests.rs
- bus_debug.rs → tests/bus_debug_tests.rs (gated: cfg(feature = "debug-checks"))
- table.rs disk_spill_tests → tests/table_disk_spill_tests.rs (gated: cfg(feature = "disk-spill"))

Visibility widened:
- compute_frame_evaluation_points: fn → pub(crate)
- BusDebugTracker.{bus_filter, logs}: private → pub(crate)
- Table.mmap_backing: private → pub(crate)
- TableMmapBacking: private struct → pub(crate) struct

All 128 tests still pass; disk-spill feature adds 5 more (133 total).
Clippy clean; production build clean.
